requirements.txt
# Core dependencies
mcp==1.1.0  # Using 1.1.0 for protocol compatibility with kanban-mcp
anthropic>=0.18.0,<1.0.0
asyncio==3.4.3
aiohttp==3.9.1
python-dotenv==1.0.0

# Data handling
pydantic>=2.8.0
python-dateutil==2.8.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# Communication integrations (optional, add as needed)
# slack-sdk==3.26.1
# sendgrid==6.11.0